added function to display grocery list items
added a function that attaches a listener to a grocery list. Anytime the
grocerylist has an item added, the DOM will be updated dynamically.
Currently have a listener bound to the removal of a grocery list item
but at this time it does nothing more than logging to the console.