Updated data model for implicit root container
When a new user is created, the root container for that user must be
implicitly created and exist before the users node is written. Updated
the security rules accordingly.

Also updated the README with information about the users tree, and to
reflect the change to a single, implicit root container.